requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.5.0
opencv-python>=4.8.0
scipy>=1.10.0
tqdm>=4.65.0
streamlit>=1.24.0